backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database and vector store
sqlalchemy==2.0.23
asyncpg==0.29.0
pgvector==0.2.4
redis==5.0.1

# AWS services
boto3>=1.34.72
botocore>=1.34.72
langchain-aws==0.1.0
langchain-community==0.0.38
langchain==0.1.0

# Authentication and security
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4

# Search and retrieval
opensearch-py==2.4.0
sentence-transformers==2.2.2

# Web crawling and content extraction
playwright==1.40.0
trafilatura==1.6.4
beautifulsoup4==4.12.2
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
structlog==23.2.0
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1
